IntegerSlider min and max getters were doubles for some reason

+++ src/lab/component/swing/input/IntegerSlider.java
@@ -37,11 +37,11 @@ public class IntegerSlider extends InputComponent implements ChangeListener {
 		slider.setValue(v);
 	}
 
-	public double getMin() {
+	public int getMin() {
 		return min;
 	}
 
-	public double getMax() {
+	public int getMax() {
 		return max;
 	}
